Fix deletion screen

+++ src/main/webapp/scripts/app/webupdates/modalDeleteObjectController.js
@@ -6,15 +6,31 @@ angular.module('webUpdates').controller('ModalDeleteObjectController', [ '$scope
         $scope.reason = 'I don\'t need this object';
         $scope.objectType = objectType;
         $scope.name = name;
+        $scope.referencesInfo = undefined;
+        $scope.isPartOfSimplePair = undefined;
 
-        getReferences(source, $scope.objectType, $scope.name);
+        $scope.doDelete = doDelete;
+        $scope.doCancel = doCancel;
+        $scope.displayUrl = displayUrl;
+        $scope.primaryKey = primaryKey;
+        $scope.isEqualTo = isEqualTo;
+        $scope.canBeDeleted = canBeDeleted;
 
-        $scope.delete = function () {
-            var deleteWithRefs = $scope.isMntPersonReference($scope.referencesInfo.references);
+        _initialisePage();
+
+        function _initialisePage() {
+            getReferences(source, $scope.objectType, $scope.name);
+        }
+
+        function doDelete() {
+            if( !canBeDeleted($scope.referencesInfo)) {
+                return;
+            }
+
+            var deleteWithRefs = isSimpleCrossReference($scope.objectType, $scope.name, $scope.referencesInfo.references);
 
             RestService.deleteObject(source, $scope.objectType, $scope.name, $scope.reason, deleteWithRefs).then(
                 function (resp ) {
-                    $log.info('Successfully deleted object:' + JSON.stringify(resp) );
                     $modalInstance.close();
 
                     $state.transitionTo('deleted', {
@@ -22,20 +38,18 @@ angular.module('webUpdates').controller('ModalDeleteObjectController', [ '$scope
                         objectType: $scope.objectType,
                         name: $scope.name
                     });
-
                 },
                 function (error) {
-                    $log.error('Error deleting object:' + JSON.stringify(error) );
                     $modalInstance.dismiss(error.data);
                 }
             );
         };
 
-        $scope.cancel = function () {
+        function doCancel() {
             $modalInstance.close();
         };
 
-        $scope.displayUrl = function(ref) {
+        function displayUrl(ref) {
             return $state.href('display', {
                 source: source,
                 objectType: ref.type,
@@ -43,61 +57,144 @@ angular.module('webUpdates').controller('ModalDeleteObjectController', [ '$scope
             });
         };
 
-        $scope.primaryKey = function(ref) {
+        function primaryKey(ref) {
             var pkey = ref.primaryKey[0].value;
 
             for (var i = 1; i < ref.primaryKey.length; i++) {
-                pkey = pkey + '/' + ref.primaryKey[i].value;
+                pkey = pkey + ref.primaryKey[i].value;
             }
 
-            return pkey;
+            return pkey.toUpperCase();
         };
 
-        $scope.isMntPersonReference = function(references) {
-            if(references.length === 0) {
+        function hasReferences(referencesInfo) {
+            if( !referencesInfo || !referencesInfo.references) {
+                //$log.info("hasReferences: No reference data available yet")
+                return false;
+            }
+            if (referencesInfo.references.length === 0) {
+                //$log.info(selfName +'hasReferences: Has no references')
+                return false;
+            }
+            return true;
+        }
+
+        function isReferencedOnlyBySelf(selfType, selfName, references) {
+            // typical case: self referenced mntner
+            var selfOnly =  _.every( references, function(ref) {
+                return $scope.isEqualTo(selfType, selfName, ref);
+            });
+            //$log.info(selfName + ' isReferencedOnlyBySelf:' + selfOnly );
+            return selfOnly;
+        }
+
+        function isRefMntBySelf( selfType, selfName, ref ) {
+            var mntBySelf = false;
+            if (selfType === 'mntner') {
+                mntBySelf =  _.any( ref.attributes, function(attr) {
+                    if( attr.name === 'mnt-by' && attr.value === selfName ) {
+                        return true;
+                    }
+                });
+            }
+            //$log.info(selfName + ' isRefMntBySelf:' + mntBySelf );
+            return mntBySelf;
+        }
+
+        function isRefAdmninedBySelf(selfType, selfName, ref ) {
+            var adminedBySelf = false;
+            if (selfType === 'role' || selfType === 'person'  ) {
+                adminedBySelf = _.any( ref.attributes, function(attr) {
+                    if( attr.name === 'admin-c' && attr.value === selfName ) {
+                        return true;
+                    }
+                });
+            }
+            //$log.info(selfName + ' isRefAdmninedBySelf:' + adminedBySelf );
+            return adminedBySelf;
+        }
+
+        function isSimpleCrossReference(selfType, selfName, references) {
+            // typical case: mntner person pair
+            if( !references || references.length === 0 ) {
+                //$log.info('Not referenced' );
                 return false;
             }
 
-            if ((references.length === 1) && $scope.isItself(references[0], $scope.objectType)) {
-                $log.info('References itself')
-                return false; // wrong? 
+            if( references.length > 1 ) {
+                //$log.info('Referenced by multiple objects' );
+                return false;
             }
 
-            if($scope.objectType.toUpperCase() === 'MNTNER') {
-                return _.every(references, function(ref) {
-                    return ref.type.toUpperCase() === 'ROLE' ||
-                        ref.type.toUpperCase() === 'PERSON' ||
-                        $scope.isItself(ref, objectType);
-                });
+            // order does matter here: check for isPartOfPair before peforming other checks
+            if( !_.isUndefined($scope.isPartOfSimplePair)) {
+                //$log.info("Is part of a pair: " + $scope.isPartOfPair);
+                return $scope.isPartOfSimplePair;
             }
 
-            if($scope.objectType.toUpperCase() === 'ROLE' || $scope.objectType.toUpperCase() === 'PERSON') {
-                return _.every(references, function(ref) {
-                    return ref.type.toUpperCase() === 'MNTNER' ||
-                        $scope.isItself(ref, $scope.objectType);
-                });
+            if( isReferencedOnlyBySelf(selfType, selfName, references) ) {
+                return false;
+            }
+
+            if( isRefMntBySelf(selfType, selfName, references[0])) {
+                return true;
+            }
+
+            if( isRefAdmninedBySelf(selfType, selfName, references[0])) {
+                return true;
+            }
+
+            return false;
+        }
+
+        function isEqualTo(selfType, selfName, ref) {
+            return ref.type.toUpperCase() === selfType.toUpperCase() && $scope.primaryKey(ref) === selfName.toUpperCase();
+        }
+
+        function canBeDeleted(referencesInfo) {
+            if( hasReferences(referencesInfo) == false) {
+                return true;
+            }
+
+            if( isReferencedOnlyBySelf($scope.objectType, $scope.name, $scope.referencesInfo.references)) {
+                return true;
+            }
+
+            if( isSimpleCrossReference($scope.objectType, $scope.name, $scope.referencesInfo.references) ) {
+                return true;
             }
 
             return false;
         };
 
-        $scope.isItself = function(ref) {
-            return (ref.type.toUpperCase() === $scope.objectType.toUpperCase() && $scope.primaryKey(ref).toUpperCase() === $scope.name.toUpperCase());
-        };
-
-        $scope.canBeDeleted = function(referencesInfo) {
-            return referencesInfo.total === 0 || $scope.isMntPersonReference(referencesInfo.references, $scope.objectType);
-        };
-
         function getReferences(source, objectType, name) {
             RestService.getReferences(source, objectType, name) .then(
                 function (resp) {
-                    $log.info('Successfully fetched object references:' + JSON.stringify(resp) );
+                    if( isSimpleCrossReference( $scope.objectType, $scope.name, resp.references)) {
+                        // fetch its pair object as well
+                        var otherType = resp.references[0].type;
+                        var otherName = primaryKey(resp.references[0]);
+                        getReferencesForPeer(source, otherType, otherName);
+                    }
                     $scope.referencesInfo = resp;
                 },
                 function (error) {
-                    $log.error('Error getting object references:' + JSON.stringify(error) );
                     $modalInstance.dismiss(error.data);
+                }
+            );
+        }
+
+        function getReferencesForPeer(source, otherType, otherName) {
+            $log.info('*** Looks like a simple cross reference: check ' + otherType + '-peer ' + otherName);
+            RestService.getReferences(source, otherType,otherName).then(
+                function( resp ) {
+                    if( isSimpleCrossReference(otherType, otherName, resp.references) ) {
+                        $log.info('*** *** ' + name + ' and ' + otherName + ' are simply cross referenced');
+                        $scope.isPartOfSimplePair = true;
+                    } else {
+                        $log.info('*** *** *** ' + name + ' and ' + otherName + ' are NOT a simply cross referenced');
+                        $scope.isPartOfSimplePair = false;
+                    }
                 }
             );
         }
